+++ src/composables/usePostHog.js
@@ -1,0 +1,14 @@
+import posthog from 'posthog-js'
+
+export function usePostHog() {
+  if (!posthog.__initialized) {
+    posthog.__initialized = true
+    posthog.init('phc_MYLWM5FCKlXA33FACSWpVJNiyjSIhqXkXUg80JO3He', {
+      api_host: 'https://us.i.posthog.com',
+      person_profiles: 'identified_only'
+    })
+  }
+
+  return { posthog }
+}
+